Changes to call API when selection options change and report on lost connectivity.
Added logic to make a new API request when the meal/diet type change is applied via the bottom sheet. Added simple listener/toast to report when the app loses connectivity.

The tutorial wanted me to add some (basic) additional logic for handling when we come 'back online', but I disliked both the design and the functionality, so I'm skipping it.