Added Search Functionality
Added Search Menu item and hooked up API call to retrieve recipes via a search query. Intuitively, I think the expected behavior in this context would be for the search to run on the list of diet/meal type items previously returned via the API, but I'm holding to the tutorial for now.